🏗️refactor: Optimize memory management and update attention weight precision
- Added `torch.cuda.empty_cache()` in `TorchMoECallback` to improve memory management during training.
- Changed attention weight precision from `float32` to `bfloat16` in `g3moe_model.py` for enhanced performance and reduced memory usage.
- Refactored `main` function in `custom_model_sft.py` to accept model, data, and training configurations as parameters, improving modularity and clarity.

+++ models/g3moe_model.py
@@ -1019,7 +1019,7 @@ def eager_attention_forward(
         attn_weights = attn_weights + causal_mask
 
     # upcast attention to fp32
-    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.float32).to(query.dtype)
+    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.bfloat16).to(query.dtype)
     attn_weights = nn.functional.dropout(attn_weights, p=dropout, training=module.training)
     attn_output = torch.matmul(attn_weights, value_states)
     attn_output = attn_output.transpose(1, 2).contiguous()
